created a component directory with custom_navigation_button.dart, where  placed the CustomNavigationButton class
