+++ lib/home_page.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:udemy_flutter_course/components/custom_navigation_button.dart';
 import 'package:udemy_flutter_course/section/10/quizzler.dart';
 import 'package:udemy_flutter_course/section/11/destini.dart';
 import 'package:udemy_flutter_course/section/12/bmi_calculator.dart';
@@ -59,43 +60,6 @@ class HomePage {
             assetImage: 'images/clima.png',
           ),
         ],
-      ),
-    );
-  }
-}
-
-class CustomNavigationButton extends StatelessWidget {
-  const CustomNavigationButton({
-    required this.routName,
-    required this.assetImage,
-    Key? key,
-  }) : super(key: key);
-  final String assetImage;
-  final String routName;
-
-  @override
-  Widget build(BuildContext context) {
-    return ElevatedButton(
-      child: Image(
-        image: AssetImage(assetImage),
-      ),
-      onPressed: () {
-        Navigator.pushNamed(
-          context,
-          routName,
-        );
-      },
-      style: ButtonStyle(
-        backgroundColor: MaterialStateProperty.all<Color>(
-          Colors.transparent,
-        ),
-        shadowColor: MaterialStateProperty.all<Color>(
-          Colors.transparent,
-        ),
-        overlayColor: MaterialStateProperty.all<Color>(
-          Colors.transparent,
-        ),
-        splashFactory: NoSplash.splashFactory,
       ),
     );
   }
